Crate service_discovery

Service Discovery

Discover other instances of your application on the local network.

Structs

ServiceDiscovery
ServiceDiscovery is a RAII API for the purposes of discovering peers on the local network. To stop the process of discovery completely, it is sufficient to drop the instance of this struct.